requirements.txt
python-telegram-bot==13.12

flask==2.1.2
gunicorn==20.1.0

ipython==8.4.0
python-dotenv==0.20.0
